Reorganize frame_spec.rb to match Puppeteer's test structure
- Align frame_spec.rb with puppeteer/puppeteer test ordering
- Add missing test: should click elements in a frameset
- Extract Ruby-specific tests to frame_ext_spec.rb:
  - #execution_context tests
  - should report frame.name()
